tools/build_scripts: Add script to generate UAC manifests
We need to add this script to generate Windows manifest files during the
build to allow test programs with 'update' in the filename to run without
UAC from getting into our way, for 32-bit Windows builds.

The following commits will update the build files to embed this manifest
as appropriate, so that test programs can run properly on 32-bit Windows
builds.